+++ PID_Py/__init__.py
@@ -1,4 +1,4 @@
 import importlib.metadata
 
 __version__ = importlib.metadata.version("PID_Py")
-__author__ = "ThunderTecke"
+__author__ = "ThunderTecke <thunder.tecke@gmail.com>"